fixes due to changes in dolfinx

+++ python/dolfinx_contact/one_sided/snes_against_plane.py
@@ -86,7 +86,7 @@ def snes_solver(mesh: dmesh.Mesh, mesh_data: Tuple[dmesh.MeshTags, int, int],
     v = ufl.TestFunction(V)
     dx = ufl.Measure("dx", domain=mesh)
     F = ufl.inner(sigma(u), epsilon(v)) * dx - \
-        ufl.inner(_fem.Constant(mesh, [0, ] * mesh.geometry.dim), v) * dx
+        ufl.inner(_fem.Constant(mesh, np.float64([0, ] * mesh.geometry.dim)), v) * dx
 
     # Stored strain energy density (linear elasticity model)    # penalty = 0
     # psi = 1/2*ufl.inner(sigma(u), epsilon(u))
